refactor(main): extract host registry into host/registry
Pulls ComfyWindowEntry, ComfyPanelKey, BodyMode, VALID_PANELS, comfyWindows, the installationIdToWindowKey secondary index, pendingAttachClaims, lastFocusedInstallationId, and the find/openOrFocus helpers (plus bringToFront) out of index.ts and into src/main/host/registry.ts.

Late-bound chooser-host construction via setHostFactories({ createChooser }) so the registry never imports host-construction code (no cycle). index.ts wires the factory at the top of whenReady.

index.ts 3720 -> 3125 lines (-595).

Amp-Thread-ID: https://ampcode.com/threads/T-019e2090-2750-71d1-a4c9-c5aab842cc37
